[backend] - (userOrganization) remove user from user_organization_pending table when adding to user_organization #470
